Fixed issues with user validation when editing or deleting

+++ controllers/api/product-routes.js
@@ -3,6 +3,10 @@ const { Product, Category, Registry } = require("../../models");
 const router = require("express").Router();
 
 router.post("/", async (req, res) => {
+  if (!req.session.loggedIn) {
+    return;
+  }
+
   console.log("Received post req to create product in db");
   try {
     /* check if category exists */
@@ -10,17 +14,31 @@ router.post("/", async (req, res) => {
     if (!categoryData) {
       res.status(404).json("Category not found");
     }
+
     /* check if registry exists */
-    const registryData = Registry.findByPk(req.body.registry_id);
+    const registryData = await Registry.findOne({
+      where: {
+        id: req.body.registry_id,
+        user_id: req.session.userId,
+      },
+    });
+
     if (!registryData) {
-      res.status(404).json("Registry not found");
+      res.status(404).json({ message: "Registry not found" });
+      return;
     }
+
     /* all ok, create the product */
-    const productData = await Product.create(req.body);
+    const productData = await Product.create({
+      ...req.body,
+      user_id: req.session.userId,
+    });
+
     if (productData) {
       res.status(200).json(productData);
     } else {
       res.status(404).json("Product not created");
+      return;
     }
   } catch (err) {
     let error = "Error creating product";
@@ -42,7 +60,12 @@ router.post("/", async (req, res) => {
     res.status(500).json({ success: false, message: error });
   }
 });
+
 router.put("/:id", async (req, res) => {
+  if (!req.session.loggedIn) {
+    return;
+  }
+
   console.log("Received put req to update product in db");
   try {
     /* check if category exists */
@@ -52,7 +75,12 @@ router.put("/:id", async (req, res) => {
       return;
     }
     /* check if registry exists */
-    const registryData = await Registry.findByPk(req.body.registry_id);
+    const registryData = await Registry.findOne({
+      where: {
+        id: req.body.registry_id,
+        user_id: req.session.userId,
+      },
+    });
     if (!registryData) {
       res.status(404).json("Registry not found");
       return;
@@ -80,17 +108,23 @@ router.put("/:id", async (req, res) => {
 });
 
 router.delete("/:id", async (req, res) => {
+  if (!req.session.loggedIn) {
+    return;
+  }
+
   console.log("received delete req to /api/product/:id" + req.params.id);
   try {
     const product = await Product.destroy({
       where: {
         id: req.params.id,
+        user_id: req.session.userId,
       },
     });
     if (product) {
       res.status(200).json({ message: "Product deleted successfully" });
     } else {
-      res.status(400).json({ message: "Product not deleted" });
+      res.status(404).json({ message: "Product not found" });
+      return;
     }
   } catch (err) {
     console.log(err);
